Add useSetupSocketHandlers() hook
